Scenery for both water and terrain done.

+++ Terrain.cpp
@@ -14,7 +14,7 @@ using namespace std;
 #define SCENE_MODE 0
 
 /* Flat Terrain. Ability to input a height map: either real or generated from different applications. Shader that adds at least 3 different type of terrain(grass, desert, snow). */
-Terrain::Terrain(float x_d, float z_d, const char* terrain_0, const char* terrain_1, const char* terrain_2, const char* terrain_3, const char* blend_map)
+Terrain::Terrain(int x_d, int z_d, const char* terrain_0, const char* terrain_1, const char* terrain_2, const char* terrain_3, const char* blend_map)
 {
 	//Setup the terrain.
 	this->x = x_d * SIZE;
@@ -31,7 +31,7 @@ Terrain::Terrain(float x_d, float z_d, const char* terrain_0, const char* terrai
 }
 
 /* Procedurally generated Terrain. Ability to input a height map: either real or generated from different applications. Shader that adds at least 3 different type of terrain(grass, desert, snow). */
-Terrain::Terrain(float x_d, float z_d, const char* terrain_0, const char* terrain_1, const char* terrain_2, const char* terrain_3, const char* blend_map, const char* height_map)
+Terrain::Terrain(int x_d, int z_d, const char* terrain_0, const char* terrain_1, const char* terrain_2, const char* terrain_3, const char* blend_map, const char* height_map)
 {
 	//Setup the terrain.
 	this->x = x_d * SIZE;
